Add files via upload
The First The report analyzes the impact of Gaussian distribution on system models through seven tasks split into three parts. First, we examine the accuracy of an estimator based on sequence length affected by Gaussian distribution. Secondly, we focus on adding and subtracting two Gaussian random variables. In addition, we examine the periodogram and noise correlation of two data sequences of a system affected by Gaussian noise. Finally, we are involved in deriving and plotting the PSD and AFC of the input and output signal of an autoregressive (AR) process.

The second report delves into the significance of the receiver in a wireless digital communication system, precisely its equalizer and decoder. It is imperative to have a decoding key in place to prevent distortions during transmission and ensure that clear images are received through wireless channels. The recipient needs to have suitable equalizer and detector models to ensure the precise decoding of the encrypted image. The primary aim is to generate an unencrypted image that closely resembles the original one. We elaborate on selecting the appropriate filter order for the received signal, enabling us to determine the filter coefficients and reconstruct the disrupted key.